Change double to long to communicate issue more clearly

+++ tests/MongoDB.Driver.Tests/Linq/Linq3Implementation/Jira/OData25Tests.cs
@@ -29,7 +29,7 @@ public class OData25Tests : Linq3IntegrationTest
 
         // conversion fails
         var queryable = collection.AsQueryable()
-            .Select(c => (int?)((double)c.x * 100));
+            .Select(c => (int?)((long)c.x * 100));
 
         var result = queryable.ToList();
         result[0].Should().Be(100);
@@ -55,7 +55,7 @@ public class OData25Tests : Linq3IntegrationTest
         var collection = GetCollection();
 
         var queryable = collection.AsQueryable()
-            .Select(c => (double?)((int)c.x * 100));
+            .Select(c => (long?)((int)c.x * 100));
 
         var result = queryable.ToList();
         result[0].Should().Be(100);
